引入 MyBatis-Plus 并完善 engine 层
- 引入 MyBatis-Plus 3.5.16（spring-boot4-starter + jsqlparser 分页插件），
  双数据源改用 MybatisSqlSessionFactoryBean，挂载 PostgreSQL 分页插件
- UserMapper 改为继承 BaseMapper，updateById 默认跳过 null 字段实现部分更新
- DbEntity 加 @TableId(IdType.AUTO) / @TableField(updateStrategy=NEVER) 保护创建时间
- UserEntity 加 @TableName，数值状态字段改用包装类（null=未设置）
- 新增 engine 容器/事件/注册器工具类（UserContainer、DataContainer、
  BaseHandlerRegister、UserInitFinishListener、SpringUtils 等）

+++ src/main/java/com/ytrue/game/framework/database/data/entity/WanderSubtitleEntity.java
@@ -8,7 +8,7 @@ import java.io.Serial;
 import java.util.Date;
 
 /**
- * 游走字幕实体（表 {@code tbl_app_wander_subtitle}）。
+ * 游走字幕实体（表 {@code app_wander_subtitle}）。
  *
  * @since 1.0.0
  */
@@ -27,7 +27,7 @@ public class WanderSubtitleEntity extends DbEntity {
     /**
      * 间隔时间（分钟）
      */
-    private int intervalTime;
+    private Integer intervalTime;
 
     /**
      * 开始时间
@@ -42,6 +42,6 @@ public class WanderSubtitleEntity extends DbEntity {
     /**
      * 下次推送时间
      */
-    private long nextSendTime;
+    private Integer nextSendTime;
 
 }
